+++ sentry-android-replay/src/main/java/io/sentry/android/replay/video/SimpleVideoEncoder.kt
@@ -33,7 +33,9 @@ import android.annotation.TargetApi
 import android.graphics.Bitmap
 import android.media.MediaCodec
 import android.media.MediaCodecInfo
+import android.media.MediaCodecList
 import android.media.MediaFormat
+import android.os.Build
 import android.view.Surface
 import io.sentry.SentryLevel.DEBUG
 import io.sentry.SentryOptions
@@ -50,8 +52,22 @@ internal class SimpleVideoEncoder(
     val onClose: (() -> Unit)? = null
 ) {
 
+    private val hasExynosCodec: Boolean by lazy(NONE) {
+        // MediaCodecList ctor will initialize an internal in-memory static cache of codecs, so this
+        // call is only expensive the first time
+        MediaCodecList(MediaCodecList.REGULAR_CODECS)
+            .codecInfos
+            .any { it.name.contains("c2.exynos") }
+    }
+
     internal val mediaCodec: MediaCodec = run {
-        val codec = MediaCodec.createEncoderByType(muxerConfig.mimeType)
+        // c2.exynos.h264.encoder seems to have problems encoding the video (Pixel and Samsung devices)
+        // so we use the default encoder instead
+        val codec = if (hasExynosCodec) {
+            MediaCodec.createByCodecName("c2.android.avc.encoder")
+        } else {
+            MediaCodec.createEncoderByType(muxerConfig.mimeType)
+        }
 
         codec
     }
@@ -139,10 +155,13 @@ internal class SimpleVideoEncoder(
     }
 
     fun encode(image: Bitmap) {
-        // NOTE do not use `lockCanvas` like what is done in bitmap2video
-        // This is because https://developer.android.com/reference/android/media/MediaCodec#createInputSurface()
-        // says that, "Surface.lockCanvas(android.graphics.Rect) may fail or produce unexpected results."
-        val canvas = surface?.lockHardwareCanvas()
+        // it seems that Xiaomi devices have problems with hardware canvas, so we have to use
+        // lockCanvas instead https://stackoverflow.com/a/73520742
+        val canvas = if (Build.MANUFACTURER.contains("xiaomi", ignoreCase = true)) {
+            surface?.lockCanvas(null)
+        } else {
+            surface?.lockHardwareCanvas()
+        }
         canvas?.drawBitmap(image, 0f, 0f, null)
         surface?.unlockCanvasAndPost(canvas)
         drainCodec(false)